requirements.txt
apache-airflow[celery]==2.10.5 --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.5/constraints-3.9.txt"
apache-airflow-providers-common-sql
apache-airflow-providers-fab
apache-airflow-providers-http
apache-airflow-providers-mysql
boto3
deprecated
duckdb
google-generativeai
graphviz
numpy
openpyxl
pandas
pendulum
playwright
pretty_errors
pymysql
pytest
pyarrow
python-dotenv
pytz
requests
sqlalchemy==1.4.54 #this version is compatible with airflow 3.0.0, specially flask-appbuilder 4.5.3  whichrequires SQLAlchemy<1.5
tqdm
xlrd
